+++ client/src/App.js
@@ -9,7 +9,12 @@ import Unauthorized from './components/Unauthorized';
 import Lounge from './components/Lounge';
 import LinkPage from './components/LinkPage';
 import RequireAuth from './components/RequireAuth';
-import { Routes, Route } from 'react-router-dom';
+import { Routes, Route, useNavigate, useLocation } from 'react-router-dom';
+import { useEffect, useState } from 'react';
+import useAxiosPrivate from './hooks/useAxiosPrivate';
+import DogAdd from './components/DogAdd';
+import DogDetail from './components/DogDetail';
+import DogEdit from './components/DogEdit';
 
 const ROLES = {
   'User': 2001,
@@ -18,7 +23,45 @@ const ROLES = {
 }
 
 function App() {
+  const [dogs, setDogs] = useState([]);
+  const [url, setUrl] = useState('/dogs/?limit=3&offset=0');
+  const axiosPrivate = useAxiosPrivate();
+  const navigate = useNavigate();
+  const location = useLocation();
 
+  const getDogs = async (url, options = null) => {
+      setUrl(url);
+      try {
+          const response = await axiosPrivate.get(url, options);
+          console.log(response.data);
+          setDogs(response.data);
+      } catch (err) {
+          console.error(err);
+          navigate('/login', { state: { from: location }, replace: true });
+      }
+  }
+  useEffect(() => {
+      const controller = new AbortController();
+      getDogs(url, {
+          signal: controller.signal
+      });
+      return () => {
+          controller.abort();
+      }
+  }, []);
+
+  const dogAddHandler = async ({name}) => {
+    console.log('DOG: ', name);
+    const response = await axiosPrivate.post('/dogs/', JSON.stringify({id: 0, name}));
+    console.log(response.data);
+    getDogs(url);
+  }
+  const dogUpdateHandler = async (dog) => {
+    console.log('DOG: ', dog);
+    const response = await axiosPrivate.put('/dogs/', JSON.stringify(dog));
+    console.log(response.data);
+    getDogs(url);
+  }
   return (
     <Routes>
       <Route path="/" element={<Layout />}>
@@ -34,7 +77,10 @@ function App() {
         </Route>
 
         <Route element={<RequireAuth allowedRoles={[ROLES.Editor]} />}>
-          <Route path="dogs" element={<Dogs />} />
+          <Route path="dogs" element={<Dogs dogs={dogs} getDogs={getDogs} />} />
+          <Route path="dogs/create" element={<DogAdd addHandler={dogAddHandler} />} />
+          <Route path="dogs/view/:id" element={<DogDetail />} />
+          <Route path="dogs/edit/:id" element={<DogEdit updateHandler={dogUpdateHandler} />} />
         </Route>
