+++ client/src/components/Dogs.js
@@ -1,33 +1,7 @@
-import { useState, useEffect } from "react";
-import useAxiosPrivate from "../hooks/useAxiosPrivate";
-import { useNavigate, useLocation } from "react-router-dom";
+import { Link } from 'react-router-dom';
+import DogCard from "./DogCard";
 
-const Dogs = () => {
-    const [dogs, setDogs] = useState();
-    const axiosPrivate = useAxiosPrivate();
-    const navigate = useNavigate();
-    const location = useLocation();
-
-    const getDogs = async (url, options) => {
-        try {
-            const response = await axiosPrivate.get(url, options);
-            console.log(response.data);
-            setDogs(response.data);
-        } catch (err) {
-            console.error(err);
-            navigate('/login', { state: { from: location }, replace: true });
-        }
-    }
-    useEffect(() => {
-        const controller = new AbortController();
-        getDogs('/dogs/?limit=3&offset=0', {
-            signal: controller.signal
-        });
-        return () => {
-            controller.abort();
-        }
-    }, []);
-
+const Dogs = ({dogs, getDogs}) => {
     const paginationHandler = (e) => {
         e.preventDefault();
         const name = e.target.getAttribute('data-name');
@@ -38,11 +12,11 @@ const Dogs = () => {
     }
     return (
         <article>
-            <h2>Dogs List</h2>
+            <h2>Dogs List (<Link to="/dogs/create">Create</Link>)</h2>
             {dogs?.data?.length
                 ? (
                     <>
-                    <table border="1" cellpading="5" cellspacing="5">
+                    <table border="1" cellpading="5" cellSpacing="5">
                         <thead>
                             <tr>
                                 <th>ID</th>
@@ -53,15 +27,7 @@ const Dogs = () => {
                         <tbody>
                     {
                         dogs.data.map((dog, i) =>
-                            <tr key={dog.id}>
-                                <td>{dog.id}</td>
-                                <td>{dog.name}</td>
-                                <td>
-                                    <a href=""> View </a> |
-                                    <a href=""> Edit </a> |
-                                    <a href=""> Delete </a>
-                                </td>
-                            </tr>
+                            <DogCard dog={dog} />
                         )
                     }
                         </tbody>
