+++ server/src/v1/dog/test.http
@@ -1,4 +1,4 @@
-@host = http://127.0.0.1:6000/v1/dogs
+@host = http://127.0.0.1:8080/v1/dogs
 
 
 ### Get all dogs
@@ -6,12 +6,12 @@ GET {{host}}/?fields=name&offset=0&limit=2&token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI
 
 
 ### Create new dogs
-POST {{host}}/?token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VybmFtZSI6ImFkbWluIiwiaWQiOjEwMCwiZXhwIjoxNjUyODUyMDE0fQ.e7fPAyd3kO-0s_QRoUVpR9KcbbjW24e8-7DKBeGv6WI HTTP/1.1
+POST {{host}}/?token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VybmFtZSI6ImFkbWluIiwiaWQiOjEwMCwiZXhwIjoxNjUzNzEwNzM2fQ.BiHhHGRIAm7pqko8cTsVgyraftw0iCSmiw3W8Y3Tm5U HTTP/1.1
 Content-Type: application/json
 
 {
     "id": 0,
-    "name": "Blackie5"
+    "name": "123"
 }
 
 
@@ -26,10 +26,10 @@ Content-Type: application/json
 
 
 ### Delete dogs
-DELETE {{host}}/ HTTP/1.1
+DELETE {{host}}/?token=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VybmFtZSI6ImFkbWluIiwiaWQiOjEwMCwiZXhwIjoxNjUzNjc0OTI3fQ.--EvNx1sqeODptZwkISe1_tB9j_81WfjNF0gFYkBuWw HTTP/1.1
 Content-Type: application/json
 
-[21,23,12,16,17]
+[49]
